+++ apps/cli/src/adapters/commander/commands/init.ts
@@ -1,27 +1,29 @@
-import { Answers } from "@pagopa/monorepo-generator";
-import loadMonorepoScaffolder, {
-  answersSchema,
+import { setMonorepoGenerator } from "@pagopa/monorepo-generator";
+import {
+  Payload as Answers,
+  payloadSchema as answersSchema,
   PLOP_MONOREPO_GENERATOR_NAME,
-} from "@pagopa/monorepo-generator";
+} from "@pagopa/monorepo-generator/generators/monorepo";
 import chalk from "chalk";
 import { Command } from "commander";
 import { $ } from "execa";
-import { okAsync, ResultAsync } from "neverthrow";
+import { errAsync, okAsync, ResultAsync } from "neverthrow";
 import { PlopGenerator } from "node-plop";
 import * as path from "node:path";
-import { Octokit } from "octokit";
 import { oraPromise } from "ora";
 
+import {
+  GitHubService,
+  PullRequest,
+  Repository,
+  RepositoryNotFoundError,
+} from "../../../domain/github.js";
 import { tf$ } from "../../execa/terraform.js";
 import { getGenerator, getPrompts, initPlop } from "../../plop/index.js";
 import { decode } from "../../zod/index.js";
 import { exitWithError } from "../index.js";
 
 type InitResult = {
-  csp: {
-    location: string;
-    name: string;
-  };
   pr?: PullRequest;
   repository?: Repository;
 };
@@ -31,29 +33,10 @@ type LocalWorkspace = {
   repository: Repository;
 };
 
-type PullRequest = {
-  url: string;
-};
-
 type RepositoryPullRequest = {
   pr?: PullRequest;
   repository: Repository;
 };
-
-class Repository {
-  get ssh(): string {
-    return `git@github.com:${this.owner}/${this.name}.git`;
-  }
-
-  get url(): string {
-    return `https://github.com/${this.owner}/${this.name}`;
-  }
-
-  constructor(
-    public readonly name: string,
-    public readonly owner: string,
-  ) {}
-}
 
 const withSpinner = <T>(
   text: string,
@@ -73,12 +56,27 @@ const withSpinner = <T>(
     },
   );
 
-// TODO: Check repository already exists: if exists, return an error
 // TODO: Check Cloud Environment exists
 // TODO: Check CSP CLI is installed
 // TODO: Check user has permissions to handle Terraform state
-const validateAnswers = (answers: Answers): ResultAsync<Answers, Error> =>
-  okAsync(answers);
+const validateAnswers =
+  (githubService: GitHubService) =>
+  (answers: Answers): ResultAsync<Answers, Error> =>
+    ResultAsync.fromPromise(
+      githubService.getRepository(answers.repoOwner, answers.repoName),
+      (error) => error as Error,
+    )
+      .andThen(({ fullName }) =>
+        errAsync(new Error(`Repository ${fullName} already exists.`)),
+      )
+      .orElse((error) =>
+        error instanceof RepositoryNotFoundError
+          ? // If repository is not found, it's safe to proceed
+            okAsync(answers)
+          : // Otherwise, propagate the error
+            errAsync(error),
+      )
+      .map(() => answers);
 
 const runGeneratorActions = (generator: PlopGenerator) => (answers: Answers) =>
   withSpinner(
@@ -89,10 +87,8 @@ const runGeneratorActions = (generator: PlopGenerator) => (answers: Answers) =>
   ).map(() => answers);
 
 const displaySummary = (initResult: InitResult) => {
-  const { csp, pr, repository } = initResult;
+  const { pr, repository } = initResult;
   console.log(chalk.green.bold("\nWorkspace created successfully!"));
-  console.log(`- Cloud Service Provider: ${chalk.cyan(csp.name)}`);
-  console.log(`- CSP location: ${chalk.cyan(csp.location)}`);
 
   if (repository) {
     console.log(`- Name: ${chalk.cyan(repository.name)}`);
@@ -182,12 +178,12 @@ const initializeGitRepository = (repository: Repository) => {
 };
 
 const handleNewGitHubRepository =
-  (octokit: Octokit) =>
+  (githubService: GitHubService) =>
   (answers: Answers): ResultAsync<RepositoryPullRequest, Error> =>
     createRemoteRepository(answers)
       .andThen(initializeGitRepository)
       .andThen((localWorkspace) =>
-        createPullRequest(octokit)(localWorkspace).map((pr) => ({
+        createPullRequest(githubService)(localWorkspace).map((pr) => ({
           pr,
           repository: localWorkspace.repository,
         })),
@@ -196,21 +192,13 @@ const handleNewGitHubRepository =
 const makeInitResult = (
   answers: Answers,
   { pr, repository }: RepositoryPullRequest,
-): InitResult => {
-  const csp = {
-    location:
-      answers.csp === "azure" ? answers.azureLocation : answers.awsRegion,
-    name: answers.csp,
-  };
-  return {
-    csp,
-    pr,
-    repository,
-  };
-};
+): InitResult => ({
+  pr,
+  repository,
+});
 
 const createPullRequest =
-  (octokit: Octokit) =>
+  (githubService: GitHubService) =>
   ({
     branchName,
     repository,
@@ -219,7 +207,7 @@ const createPullRequest =
       "Creating Pull Request...",
       "Pull Request created successfully!",
       "Failed to create Pull Request.",
-      octokit.rest.pulls.create({
+      githubService.createPullRequest({
         base: "main",
         body: "This PR contains the scaffolded monorepo structure.",
         head: branchName,
@@ -228,16 +216,15 @@ const createPullRequest =
         title: "Scaffold repository",
       }),
     )
-      .map(({ data }) => ({ url: data.html_url }))
       // If PR creation fails, don't block the workflow
       .orElse(() => okAsync(undefined));
 
 type InitCommandDependencies = {
-  octokit: Octokit;
+  gitHubService: GitHubService;
 };
 
 export const makeInitCommand = ({
-  octokit,
+  gitHubService,
 }: InitCommandDependencies): Command =>
   new Command()
     .name("init")
@@ -250,7 +237,7 @@ export const makeInitCommand = ({
         .action(async function () {
           await checkPreconditions()
             .andThen(initPlop)
-            .andTee(loadMonorepoScaffolder)
+            .andTee(setMonorepoGenerator)
             .andThen((plop) => getGenerator(plop)(PLOP_MONOREPO_GENERATOR_NAME))
             .andThen((generator) =>
               // Ask the user the questions defined in the plop generator
@@ -258,12 +245,12 @@ export const makeInitCommand = ({
                 // Decode the answers to match the Answers schema
                 .andThen(decode(answersSchema))
                 // Validate the answers (like checking permissions, checking GitHub user or org existence, etc.)
-                .andThen(validateAnswers)
+                .andThen(validateAnswers(gitHubService))
                 // Run the generator with the provided answers (this will create the files locally)
                 .andThen(runGeneratorActions(generator)),
             )
             .andThen((answers) =>
-              handleNewGitHubRepository(octokit)(answers).map((repoPr) =>
+              handleNewGitHubRepository(gitHubService)(answers).map((repoPr) =>
                 makeInitResult(answers, repoPr),
               ),
             )