+++ infra/modules/azure_cdn/variables.tf
@@ -65,8 +65,8 @@ variable "custom_domains" {
 variable "diagnostic_settings" {
   type = object({
     enabled                                   = bool
-    log_analytics_workspace_id                = string
-    diagnostic_setting_destination_storage_id = string
+    log_analytics_workspace_id                = optional(string)
+    diagnostic_setting_destination_storage_id = optional(string)
   })
   default = {
     enabled                                   = false
@@ -75,19 +75,18 @@ variable "diagnostic_settings" {
   }
   description = <<-EOT
     Define if diagnostic settings should be enabled.
-    if it is:
-    Specifies the ID of a Log Analytics Workspace where Diagnostics Data should be sent and
-    the ID of the Storage Account where logs should be sent. (Changing this forces a new resource to be created)
+    If enabled, specifies the ID of a Log Analytics Workspace where Diagnostics Data should be sent and
+    optionally the ID of the Storage Account where logs should be sent.
   EOT
 
   validation {
     condition = (
-      !(var.diagnostic_settings.enabled)
+      !var.diagnostic_settings.enabled
       || (
         var.diagnostic_settings.log_analytics_workspace_id != null
-        && var.diagnostic_settings.diagnostic_setting_destination_storage_id != null
+        || var.diagnostic_settings.diagnostic_setting_destination_storage_id != null
       )
     )
-    error_message = "log_analytics_workspace_id and diagnostic_setting_destination_storage_id are mandatory if diagnostic is enabled."
+    error_message = "At least one of log_analytics_workspace_id or diagnostic_setting_destination_storage_id must be specified when diagnostic settings are enabled."
   }
 }