+++ infra/modules/azure_container_app_environment/examples/complete/main.tf
@@ -32,7 +32,8 @@ data "azurerm_log_analytics_workspace" "common" {
 }
 
 module "container_app_environment" {
-  source = "../../"
+  source  = "pagopa-dx/azure-container-app-environment/azurerm"
+  version = "~> 1.0"
 
   environment         = local.environment
   resource_group_name = azurerm_resource_group.example.name