+++ infra/modules/azure_function_app/outputs.tf
@@ -44,3 +44,10 @@ output "function_app" {
     }
   }
 }
+
+output "diagnostic_settings" {
+  description = "Details of the diagnostic settings configured for the Function App."
+  value = {
+    id = try(azurerm_monitor_diagnostic_setting.this[0].id, null)
+  }
+}