+++ infra/modules/azure_role_assignments/examples/service_bus/main.tf
@@ -4,7 +4,9 @@ resource "azurerm_resource_group" "example" {
 }
 
 module "app_service_exposed" {
-  source      = "../../../azure_app_service_exposed"
+  source  = "pagopa-dx/azure-function-app-exposed/azurerm"
+  version = "~> 2.0"
+
   environment = local.environment
 
   resource_group_name = azurerm_resource_group.example.name
@@ -17,7 +19,9 @@ module "app_service_exposed" {
 }
 
 module "service_bus" {
-  source              = "../../../azure_service_bus_namespace"
+  source  = "pagopa-dx/azure-service-bus-namespace/azurerm"
+  version = "~> 0.0"
+
   environment         = local.environment
   resource_group_name = azurerm_resource_group.example.name
 
@@ -54,7 +58,9 @@ resource "azurerm_servicebus_subscription" "example2" {
 }
 
 module "roles" {
-  source          = "../../"
+  source  = "pagopa-dx/azure-role-assignments/azurerm"
+  version = "~> 1.3"
+
   principal_id    = module.app_service_exposed.app_service.app_service.principal_id
   subscription_id = data.azurerm_subscription.current.subscription_id