+++ actions/convert-mermaid-svg/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pagopa/convert-mermaid-svg
 
+## 0.1.0
+
+### Minor Changes
+
+- 4f526a2: Add support for .mmd to .svg diagram conversion, and improve mmdc command with icons support
+
 ## 0.0.1
 
 ### Patch Changes