+++ infra/modules/azure_app_configuration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # azure_app_configuration
 
+## 0.1.0
+
+### Minor Changes
+
+- e8c90c1: Allow setting of diagnostic settings
+
 ## 0.0.1
 
 ### Patch Changes