+++ infra/modules/azure_cosmos_account/CHANGELOG.md
@@ -1,5 +1,11 @@
 # azure_cosmos_account
 
+## 0.4.0
+
+### Minor Changes
+
+- 62dd211: Add Diagnostic Settings configuration, update tests, align example
+
 ## 0.3.1
 
 ### Patch Changes